+++ lib/swaggerValidator/modelValidator.ts
@@ -390,6 +390,22 @@ export class SwaggerExampleValidator {
         // todo skip url encoding
         pathParameters[parameter.name] = parameterValue;
       } else if (location === "query") {
+        // validate the api version value
+        if (parameter.name === "api-version" && parameterValue !== this.swagger.info.version) {
+          const meta = getOavErrorMeta("INVALID_REQUEST_PARAMETER", {
+            parameterName: "api-version",
+            apiVersion: parameterValue,
+          });
+          this.addErrorsFromErrorCode(
+            operation.operationId!,
+            exampleFileUrl,
+            meta,
+            operation,
+            undefined,
+            p
+          );
+          break;
+        }
         queryParameters[parameter.name] = parameterValue;
       } else if (location === "body") {
         if ((parameter as BodyParameter).schema?.format === "file") {